requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# SLM Project Requirements
# ─────────────────────────────────────────────────────────────────────────────
# Install PyTorch manually first (version must match your CUDA driver):
#   pip install torch==2.1.2 --index-url https://download.pytorch.org/whl/cu121
#
# Then:
#   pip install -r requirements.txt
#
# Python: 3.10+  |  CUDA: 12.1+  |  GPU: A100 / A6000
# ─────────────────────────────────────────────────────────────────────────────

# ── Core NeMo Stack ───────────────────────────────────────────────────────────
nemo_toolkit[all]==2.0.0
nemo-aligner==0.4.0

# ── NeMo Curator ─────────────────────────────────────────────────────────────
nemo-curator==0.5.0
dask[distributed]==2024.4.1
dask-cuda==24.4.0                   # optional: GPU-accelerated curator ops
cudf-cu12==24.4.0                   # optional: GPU dataframes (SemDedup)

# ── Data Extraction ───────────────────────────────────────────────────────────
trafilatura==1.9.0                  # HTML → clean text
warcio==1.7.4                       # WARC file parsing
beautifulsoup4==4.12.3
lxml==5.2.1

# ── Language Detection & Quality Filtering ────────────────────────────────────
fasttext==0.9.2                     # language ID + quality classifier
langdetect==1.0.9

# ── Deduplication ─────────────────────────────────────────────────────────────
datasketch==1.6.4                   # MinHash + LSH

# ── Tokenizer ─────────────────────────────────────────────────────────────────
sentencepiece==0.2.0

# ── Datasets & HuggingFace ────────────────────────────────────────────────────
datasets==2.19.1
huggingface-hub==0.23.0
transformers==4.41.0

# ── Training Utilities ────────────────────────────────────────────────────────
pytorch-lightning==2.2.4
omegaconf==2.3.0
hydra-core==1.3.2
einops==0.8.0
# apex — install from source (NVIDIA fused optimizers):
#   pip install git+https://github.com/NVIDIA/apex.git

# ── Numerics ─────────────────────────────────────────────────────────────────
numpy==1.26.4
scipy==1.13.0

# ── Evaluation ────────────────────────────────────────────────────────────────
lm-eval==0.4.3                      # EleutherAI eval harness

# ── Logging & Monitoring ──────────────────────────────────────────────────────
wandb==0.17.0
tensorboard==2.16.2

# ── AWS ───────────────────────────────────────────────────────────────────────
boto3==1.34.101
awscli==1.32.101

# ── Utilities ─────────────────────────────────────────────────────────────────
pyyaml==6.0.1
tqdm==4.66.4
regex==2024.4.16
psutil==5.9.8
